### 2023-12-26 #6172 @editorialbot
[PRE REVIEW]: KerrGeoPy: A Python Package for Computing Timelike Geodesics in Kerr Spacetime
**Submitting author:** @syp2001 (<a href="http://orcid.org/0009-0002-1152-9324">Seyong Park</a>)
**Repository:** https://github.com/BlackHolePerturbationToolkit/KerrGeoPy
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.9.1
