### 2024-03-14 #6172 @xuanxu
:wave: @duetosymmetry, @duncanmmacleod would you be willing to review this paper? I think your past experience with software packages for gravitational waves and black holes would come handy here.
