### 2024-04-08 #6587 @xuanxu
👋 Hi @Uddiptaatwork and @sterinaldi, and thank you for agreeing to review this submission for _KerrGeoPy_ !

The review will take place in this issue, and you can generate your individual reviewer checklists by asking editorialbot directly with `@editorialbot generate my checklist`

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention openjournals/joss-reviews#6587 so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

As you go over the submission, please check any items on your reviewer checklist that you feel have been satisfied. If you aren't sure how to get started, please see the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) -- and of course, feel free to ping me ([@xuanxu](https://github.com/xuanxu)) with any questions !

We aim for reviews to be completed within four weeks, or six weeks at latest. Please let me know if either of you require some more time. We can also use editorialbot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

If you any questions or concerns arise, please feel free to ask here or via email. And thank you again !
