### 2024-04-26 #6587 @sterinaldi
Hi @xuanxu, sorry for the late reply, I had a few hectic days. So far, I managed to install the code on a fresh environment, run the automated tests and played a bit with the code, which seems to work well so far. I think I'll be able to have a more thorough look at the documentation and the paper by the end of the week.
