### 2024-05-08 #6587 @syp2001
Hi @sterinaldi, sorry for the late reply. There are currently no community guidelines specified, but I am working on writing them and they should hopefully be up within the next few days.
