### 2024-05-08 #6587 @sterinaldi
Thanks! I found this template pretty useful, myself: https://github.com/nayafia/contributing-template
