### 2024-05-21 #6587 @syp2001
Hi @sterinaldi, sorry again for the delay. I finally had time to write the guidelines and they are up at https://github.com/BlackHolePerturbationToolkit/KerrGeoPy/blob/main/CONTRIBUTING.md now.
