### 2024-05-23 #6587 @Uddiptaatwork
Apologies for the delay, we were going through a few closures at our institute due to extraneous (geopolitical) reasons.

I just finished going through the paper, documentation and currently running tests for a final sanity check.
The paper itself and the documentation looks great and definitely up to par. The installation on Linux and MacOS(intel) was seamless.

I will get back to you in an hour with the sanity checks.
