### 2024-05-23 #6587 @Uddiptaatwork
- I could easily reproduce all the plots from the examples in the repo README without any errors/warnings.
- All tests passed using `pytest` on py311.

I recommend this paper for publication and would like to congratulate the authors for their great work!
Thank you.
