### 2024-05-23 #6587 @xuanxu
@syp2001 looks like we're close to being done here. Please merge [this PR](https://github.com/BlackHolePerturbationToolkit/KerrGeoPy/pull/8) and give your own paper a final proof reading.
