### 2024-05-28 3c404a2
changed license to MIT and updated version number
