### 2024-06-01 d4954d4
Small edits for JOSS review
