+++ paper/paper.bib
@@ -205,7 +205,7 @@ ADS Bibcode: 2019BAAS...51g..77T},
 @article{drasco,
        author = {{Drasco}, Steve and {Hughes}, Scott A.},
         title = "{Gravitational wave snapshots of generic extreme mass ratio inspirals}",
-      journal = {\prd},
+      journal = {Physical Review D},
      keywords = {04.70.-s, 97.60.Lf, Physics of black holes, Black holes, General Relativity and Quantum Cosmology, Astrophysics},
          year = 2006,
         month = jan,
